moved logging logic to launcher
